Ongoing coding

+++ HWInterface/RD53FWInterface.h
@@ -139,10 +139,10 @@ namespace Ph2_HwInterface
       uint16_t evtStatus;
     };
 
-    static void    DecodeEvents    (const std::vector<uint32_t>& data, uint16_t& status, std::vector<RD53FWInterface::Event>& events);
-    static bool    EvtErrorHandler (uint16_t status);
-    static void    PrintEvents     (const std::vector<RD53FWInterface::Event>& events, const std::vector<uint32_t>& pData = {});
-    static uint8_t lane2chipId     (const BeBoard* pBoard, uint16_t module_id, uint16_t chip_lane)
+    static void DecodeEvents    (const std::vector<uint32_t>& data, uint16_t& status, std::vector<RD53FWInterface::Event>& events);
+    static bool EvtErrorHandler (uint16_t status);
+    static void PrintEvents     (const std::vector<RD53FWInterface::Event>& events, const std::vector<uint32_t>& pData = {});
+    static int  lane2chipId     (const BeBoard* pBoard, uint16_t module_id, uint16_t chip_lane)
     {
       // #############################
       // # Translate lane to chip ID #
